pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthgen"
version = "0.1.0"
description = "Synthetic EEG dataset generator for probabilistic source localization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pydantic>=2.0",
    "mne>=1.5",
    "zarr>=2.16",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "nibabel>=5.4.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
tvb = [
    "tvb-library>=2.8",
    "tvb-data>=2.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["synthgen*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
pythonpath = ["."]
